requirements.txt
praw
edge-tts
pytubefix
selenium
Pillow
moviepy
numpy